+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Logging - fix logging of claims
 - Deployment - support vendor outside platform's region nuvla/nuvla#110
 
 ## [6.0.7] - 2022-12-06