+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
 - Logging - fix logging of claims
+- Deployment - support vendor outside platform's region nuvla/nuvla#110
 - Module - Fix operations of resources in collection
 - Dependencies - Update to parent 6.7.12