+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Deployment - support vendor outside platform's region nuvla/nuvla#110
 - Module - Fix operations of resources in collection
 - Dependencies - Update to parent 6.7.12
+- Module - Follow customer trial not taken into account at creation time bug fix
+- Deployment - Allow vendor and users with edit right on module with price to start it without subscription
 
 ## [6.0.7] - 2022-12-06